feat: 골드 획득 업적 처리 기능 추가
- 테스트 코드 추가
- 골드 획득이 발생하는 곳에서 로그 삽입 로직 추가

+++ dailyquest_domain/src/main/java/dailyquest/log/gold/earn/entity/GoldEarnLog.kt
@@ -8,7 +8,7 @@ import jakarta.persistence.*
 @Entity
 class GoldEarnLog(
     user: UserInfo,
-    amount: Int,
+    amount: Long,
     source: GoldEarnSource
 ): CreatedTimeEntity() {
     @Column(name = "gold_earn_log_id")
@@ -24,5 +24,5 @@ class GoldEarnLog(
     val source: GoldEarnSource = source
 
     @Column(nullable = false)
-    val amount: Int = amount
+    val amount: Long = amount
 }
